requirements.txt
amqp==2.4.2
apertium-apy==0.11.5
apertium-streamparser==5.0.2
astroid==2.2.5
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.1.7
django-ckeditor==5.7.0
django-js-asset==1.2.2
flower==0.9.3
idna==2.8
isort==4.3.15
kombu==4.5.0
lazy-object-proxy==1.3.1
lxml==4.3.3
mccabe==0.6.1
numpy==1.16.2
pdfminer.six==20181108
Pillow==6.0.0
pycryptodome==3.7.3
pylint==2.3.1
python-docx==0.8.10
pytz==2018.9
pywin32==224
requests==2.21.0
six==1.12.0
sortedcontainers==2.1.0
tornado==5.1.1
typed-ast==1.3.1
urllib3==1.24.3
vine==1.3.0
virtualenv==16.4.3
wrapt==1.11.1 
